more warnings
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@15784 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/common/TetrahedralMesh.cpp
@@ -129,7 +129,7 @@ void TetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::ConstructFromMeshReader(
         if (rMeshReader.GetNumElementAttributes() > 0)
         {
             assert(rMeshReader.GetNumElementAttributes() == 1);
-            unsigned attribute_value = element_data.AttributeValue;
+            double attribute_value = element_data.AttributeValue;
             p_element->SetAttribute(attribute_value);
         }
     }
@@ -175,7 +175,7 @@ void TetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::ConstructFromMeshReader(
         if (rMeshReader.GetNumFaceAttributes() > 0)
         {
             assert(rMeshReader.GetNumFaceAttributes() == 1);
-            unsigned attribute_value = face_data.AttributeValue;
+            double attribute_value = face_data.AttributeValue;
             p_boundary_element->SetAttribute(attribute_value);
         }
     }
